Update examples Avoid shadowing 'id' in simple module as well Avoid shadowing 'id' in Connection Fix bugs in Message classes
